Move transformations from DataSet to DataSetFilters
