+++ tests/conftest.py
@@ -129,6 +129,11 @@ def hexbeam():
 
 
 @pytest.fixture
+def grid():
+    return pyvista.UnstructuredGrid(examples.hexbeamfile)
+
+
+@pytest.fixture
 def tetbeam():
     return examples.load_tetbeam()
